user-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
scikit-learn==1.6.1
numpy==2.0.0
pandas==2.2.2
scipy==1.13.1
joblib==1.4.2
aiofiles==23.2.1
apscheduler==3.10.4
